+++ packages/build-tools/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Features
 
-- **workspace-tools:** Use `pnpm publish` in the `npm-publish` executor ([48a2ad49](https://github.com/storm-software/storm-ops/commit/48a2ad49))
+- **workspace-tools:** Use `pnpm publish` in the `npm-publish` executor
+  ([48a2ad49](https://github.com/storm-software/storm-ops/commit/48a2ad49))
 
 ## 0.112.0 (2024-12-30)